Split download and update functions out of main.cpp and into download.cpp.
These functions need access to some stuff in main.cpp, so I added externs.
